Change Glyph#getUnicode return value from Integer to int.
DEVSIX-791

+++ src/main/java/com/itextpdf/highlevel/notused/appendix/CanvasLayoutMethods.java
@@ -64,7 +64,7 @@ public class CanvasLayoutMethods {
                 if (!text.get(glyphPos).hasValidUnicode()) {
                     return false;
                 }
-                int charCode = (int)text.get(glyphPos).getUnicode();
+                int charCode = text.get(glyphPos).getUnicode();
                 return (charCode < ' ' || charCode == 'i');
             }
