+++ src/main/java/com/itextpdf/highlevel/notused/appendix/DocumentLayoutMethods.java
@@ -56,7 +56,7 @@ public class DocumentLayoutMethods {
                 if (!text.get(glyphPos).hasValidUnicode()) {
                     return false;
                 }
-                int charCode = (int)text.get(glyphPos).getUnicode();
+                int charCode = text.get(glyphPos).getUnicode();
                 return (charCode < ' ' || charCode == 'i');
             }
